pyproject.toml
[project]
name = "student-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
